refactor lib js
