+++ .gitignore
@@ -1,5 +1,7 @@
 /node_modules
 uprpc-app/dist
 uprpc-app/node_modules
-Uprpc-mock/dist
+uprpc-mock/dist
 uprpc-mock/node_modules
+uprpc-lib/dist
+uprpc-lib/node_modules
